requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
attrs==25.1.0
Authlib==1.6.0
backoff==2.2.1
beautifulsoup4==4.13.4
browser-use==0.5.5
bubus==1.4.5
cachetools==5.5.2
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.3
coremltools==8.2
cryptography==45.0.5
cycler==0.12.1
Cython==3.1.2
distro==1.9.0
fastapi==0.110.0
filelock==3.17.0
fonttools==4.61.1
fsspec==2024.12.0
fuzzywuzzy==0.18.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.26.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
groq==0.30.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
langchain==0.3.26
langchain-core==0.3.69
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.6
Levenshtein==0.26.1
markdown-it-py==3.0.0
markdown_pdf==1.7
markdownify==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mcp==1.11.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
oauthlib==3.3.1
ollama==0.5.1
openai==1.97.0
opencv-python==4.12.0.88
orjson==3.11.0
packaging==24.2
patchright==1.52.5
pillow==11.2.1
playwright==1.53.0
polars==1.36.1
polars-runtime-32==1.36.1
portalocker==2.10.1
posthog==6.1.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
pyaml==25.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
PyMuPDF==1.25.3
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyparsing==3.2.3
pypdf==5.8.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.11.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.2
scipy==1.16.3
screeninfo==0.8.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.4.1
starlette==0.36.3
sympy==1.14.0
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.48.2
typing-inspection==0.4.1
typing_extensions==4.12.2
ultralytics==8.3.241
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.3.0
uuid7==0.1.0
uvicorn==0.27.1
websockets==15.0.1
zstandard==0.23.0
